requirements.txt
Flask==2.2.5
gunicorn==20.1.0
google-cloud-firestore==2.11.0
pyjwt==2.8.0
qrcode==7.4
pillow==10.0.0
flask-cors==3.0.10
